Add haproxy backend to forward packages requests to S3
The backend servers are not declared as usual because S3 has too many IP
addresses to track.

Instead, each request will resolve the IP address (cached) for the
backend and use it to continue processing the request.

This was described by Willy Tareau in [1] a GitHub discussion:

[1] https://softwarefactory-project.io/r/c/dci-infra/+/33531

Quoting:
Then if you're using it as a service instead of a server, this is where the problem stems from. You shouldn't be using a server in load-balancing form with its automatic address resolution, health checks, weights, idle pools etc which are all reset upon every change, it would be way cleaner to just resolve on the fly when the request arrives and forward the request to the IP address that results from the resolution.

For example I think that doing something approximately like this (adapted from the config manual) would be much cleaner (and wouldn't report address changes since no address changes):

resolvers mydns
  nameserver dns1 1.1.1.1
  nameserver dns8 8.8.8.8
   timeout retry   1s
   hold valid 10s
   hold nx 3s
   hold other 3s
   hold obsolete 0s

frontend frt
    bind ...
    use_backend bck if { path /blah }

backend bck
   # dedicated to service.provider.tld
   http-request do-resolve(txn.svcip,my_dns_section,ipv4) str(service.provider.tld)
   http-request set-dst var(txn.svcip)
   server svc 0.0.0.0:80  # will go to the address choosen by set-dst above

It could also make sense to add the IP address used and the source port into the logs so as to help troubleshooting connectivity issues (correlation with tcpdump, firewalls etc).

Change-Id: Iccd2f5f18fae26260ea079816f4876afa20e95d2
Depends-On: https://softwarefactory-project.io/r/c/dci-config/+/33529