Update of ca-certificates should run async
This is the first package task that will update the dnf cache and it
sometimes ends in an Ansible timeout (ansible-playbook losing track of
the ssh connection apparently).

Enabling this task to run async should help avoid this problem.

Also retry up to 3 times the installation of ca-certifiates & dnf
refresh.